Stardle Valley v1.0
A working template. Only 5 days worth of items are currently in the database, plans include expanding this to all in-game items. Each day is currently being manually reset, automation is in the works for the days to turn over at midnight just like real Wordle. Adding a feedback section in the next update.